Feat/client binary (#13)
* Example syntax

* wip: encoding + from-file

* fix: typo

* fix: encoding option as an enum

* from file

* feat: Now can read a file as binary or ascii

Before it would assume that the file content was ASCII.

* doc: Some documentation

* Adding a dry-run option

Build the message and print in the screen instead of transmitting it.

* Alternative payload from stdin

If not given as an argument, expect it from stdin, allowing cascade
commands with pipe.

* Treat payload as binary

Before it was assumed it was a String. This allows to be ready for a
binary payload. String can always be converted into binary, but not the
other way around.

* Instructions to compile it

* fix: Our new standard is to assume payload as a binary

* This is a single shot. Don't need to lock it.